This commit: made the devices facilitator testable without devices connected. implemente data list update for ao and do task bundles. added write_data methods to all.
implement [done in this commit].:
        self.taskbundle_ao.write_data()
        self.taskbundle_do.write_data()
        self.metronome.write_data()

need to make the data list into data dict. - changed mind, no need. because it follows the natural order of the sub_tasks list.

also impelemnt [done in this commit].
        self.taskbundle_ao.update_data()
        self.taskbundle_do.update_data()
        self.metronome.update_data()
need to:
check the devicesfcltr.configs_* carries name for it, that allows you to precisely update the data in the configs_list - which will translate into the data_list. [done in this commit].
1. map new configuration data to the existing subtask_ao/do_configs_list
      need to have a test to make sure only the data fields are changed.
      need to make a demo for this. [done in this commit].

work in AcquisitionFcltr.acquisition_mode1() [done in this commit].
need to get devices ready before looping [done in this commit].
may need to decouple the write_data from get_ready() methods in the TaskBundle objects for both AO and DO.  [done in this commit].
need to get_ready before mapping data, and need to only write data after chekcout single cycles.[done in this commit].

next:
should implement the following:
keep in mind: the focused process facilitator would "play the devices" through the device facilitator, that has accesses all the devices.

2. Think about how the AcquisitionFcltr (the command) would interact with the DeviceFcltr (receiver) that orchestrates all the devices. [done]

Note - always go back to the cli to look at the structure after each item. and read the notes there! and check it by running the demo for the firstcli.py

3. Then think about how to isolate the DataFacilitator out (and how to leave a "port" for future expansion...).

4. Then create a demo to perform the acquisition for a single stack.

5. after that have a new series of commits for multi stack acquisiton.

6. work on camera module for the rest of the week, perhaps...